requirements.txt
anyascii==0.2.0
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.1.13
django-allauth==0.45.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-modelcluster==5.1
django-taggit==1.5.1
django-treebeard==4.5.1
djangorestframework==3.12.4
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
html5lib==1.1
idna==2.10
l18n==2020.6.1
oauthlib==3.1.1
openpyxl==3.0.7
Pillow==8.3.1
polib==1.1.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
telepath==0.2
urllib3==1.26.6
wagtail==2.13.4
wagtail-localize==0.9.5
webencodings==0.5.1
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.4.4
xlwt==1.3.0
